% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fakestroop-package.R
\docType{package}
\name{fakestroop-package}
\alias{fakestroop}
\alias{fakestroop-package}
\title{fakestroop: Simulate Fake Stroop Data}
\description{
Simulates fake data from the Stroop Interference Task for use in teaching and benchmarking students' data skills.
}
\author{
\strong{Maintainer}: Dale Barr \email{dale.barr@glasgow.ac.uk} (\href{https://orcid.org/0000-0002-1121-4608}{ORCID})

}
\keyword{internal}
