% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-stroop.R
\name{make_stroop}
\alias{make_stroop}
\title{Simulate stroop data}
\usage{
make_stroop(
  n_subj,
  n_trials_per_colour = 5,
  stroop_colours = c("blue", "purple", "brown", "red", "green"),
  n_native = round((sample(seq(0.55, 0.65, 0.01), 1) * n_subj)),
  n_missing_lang = sample(2:5, 1),
  lang_typos = TRUE,
  prob_incongruent_err = 0.025,
  prob_voicekey_fail = 0.01,
  prob_transcript_err = 0.04,
  congruency_eff_ms = 250,
  intercept_ms = 600,
  rt_lower_bound = 400,
  random_intercepts_sd = 100,
  random_slopes_sd = 50,
  err_sd = 200
)
}
\arguments{
\item{n_subj}{Number of subjects.}

\item{n_trials_per_colour}{Number of times each colour word will repeat for a single subject.}

\item{stroop_colours}{A character vector exhaustively listing the colour words used in the study.}

\item{n_native}{Number of native speakers.}

\item{n_missing_lang}{Number of subjects who are missing values for the \code{eng_lang} variable.}

\item{lang_typos}{Whether or not there are typos in the \code{eng_lang} variable.}

\item{prob_incongruent_err}{Proportion of incongruent trials where an error is committed.}

\item{prob_voicekey_fail}{Proportion of trials where the voice key fails.}

\item{prob_transcript_err}{Proportion of colour words responses on the transcript that contain an error.}

\item{congruency_eff_ms}{Difference between average response time in milliseconds on incongruent and congruent trials. Positive values mean slower responses on incongruent trials.}

\item{intercept_ms}{Grand mean response time in milliseconds.}

\item{rt_lower_bound}{Lowest possible value for response time.}

\item{random_intercepts_sd}{Standard deviation in grand mean RT across subjects.}

\item{random_slopes_sd}{Standard deviation in the congruency effect across subjects.}

\item{err_sd}{Residual standard deviation}
}
\value{
A list object.
}
\description{
Randomly generates data corresponding to simulated performance on the Stroop interference task.
}
\details{
See the paper for details.
}
