% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-seeds.R
\name{random_seeds}
\alias{random_seeds}
\title{Create a vector with unique random seeds.}
\usage{
random_seeds(n_seeds)
}
\arguments{
\item{n_seeds}{Number of seeds to create.}
}
\value{
A vector of randomly-generated integers from 1 to \code{.Machine$integer.max}.
}
\description{
Create a vector with unique random seeds.
}
