% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-stroop.R
\name{save_stroop}
\alias{save_stroop}
\title{Save stroop data to disk}
\usage{
save_stroop(stroop, path, difficulty = "hard", overwrite = FALSE)
}
\arguments{
\item{stroop}{Source data object created using \code{\link[=make_stroop]{make_stroop()}}.}

\item{path}{Name of the path to write the data.}

\item{difficulty}{Whether the data wrangling task should be \code{"easy"} or \code{"hard"}.}

\item{overwrite}{Whether to overwrite an existing path.}
}
\value{
Returns \code{NULL} invisibly (called for its side effect only).
}
\description{
Save Stroop data to disk with varying levels of data-wrangling difficulty.
}
