% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-stroop.R
\name{solve_stroop}
\alias{solve_stroop}
\title{Run paired t-test on stroop data}
\usage{
solve_stroop(stroop, var.equal = TRUE, alpha = 0.05)
}
\arguments{
\item{stroop}{Either a raw stroop object generated by \code{\link[=make_stroop]{make_stroop()}} or pre-wrangled data generated by \code{\link[=wrangle_stroop]{wrangle_stroop()}}.}

\item{var.equal}{Whether to assume equal variances (vanilla t-test, the default) or allow for variances to be unequal.}

\item{alpha}{Alpha level for the test.}
}
\value{
Results of the t-test in a data frame from \code{\link[broom:reexports]{broom::tidy()}}.
}
\description{
Runs a paired t-test on Stroop data, performing any needed data wrangling first.
}
