% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-stroop.R
\name{wrangle_stroop}
\alias{wrangle_stroop}
\title{Wrangle stroop data into a single dataframe}
\usage{
wrangle_stroop(stroop)
}
\arguments{
\item{stroop}{Object created by \code{\link[=make_stroop]{make_stroop()}}.}
}
\value{
A data frame.
}
\description{
Wrangle raw Stroop data into a data frame with 'idealized' data.
}
