% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save-stroop.R
\name{zip_stroop}
\alias{zip_stroop}
\title{Save stroop data to zipfile}
\usage{
zip_stroop(stroop, zip_filename, difficulty, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{stroop}{Source data object created using \code{\link[=make_stroop]{make_stroop()}}.}

\item{zip_filename}{Name of the output zipfile.}

\item{difficulty}{Whether the data wrangling task should be \code{"easy"} or \code{"hard"} (passed to \code{save_stroop()}).}

\item{overwrite}{Whether to overwrite an existing file.}

\item{quiet}{Whether to suppress messages when files are saved.}
}
\value{
Returns \code{NULL} invisibly (called for its side effect only).
}
\description{
Save Stroop data to zip file with varying levels of data-wrangling difficulty.
}
